Day 9: Visit Cell instead of Point
By explicitly specifying the transform lambda return type as Cell& we
can get a proper reference view to Cells. Thus we can mutate the Cells
directly instead of fetching them separately from the grid based on
Point.

Also noticed more peculiarities with -O and views. Apparently with the |
operator results are wrong if the lambda is defined separately but
correct if it's defined inline. With function syntax (`transform(inp,
func)`) it produces correct results..